+++ octasine/src/processing_parameters/utils.rs
@@ -1,22 +1,22 @@
 
 
-pub fn map_parameter_value_to_step<T: Copy>(steps: &[T], value: f32) -> T {
+pub fn map_parameter_value_to_step<T: Copy>(steps: &[T], value: f64) -> T {
     let value = value.max(0.0).min(1.0);
     let len = steps.len();
 
-    steps[((value * len as f32) as usize).min(len - 1)]
+    steps[((value * len as f64) as usize).min(len - 1)]
 }
 
 
 pub fn map_step_to_parameter_value<T: Copy + PartialEq>(
     steps: &[T],
     step_value: T,
-) -> f32 {
+) -> f64 {
     for (index, step) in steps.iter().enumerate(){
         if *step == step_value {
-            let fraction = 1.0 / (steps.len() - 1) as f32;
+            let fraction = 1.0 / (steps.len() - 1) as f64;
 
-            return fraction * index as f32;
+            return fraction * index as f64;
         }
     }
 
@@ -25,12 +25,12 @@ pub fn map_step_to_parameter_value<T: Copy + PartialEq>(
 
 
 pub fn map_parameter_value_to_value_with_steps(
-    steps: &[f32],
-    parameter_value: f32
-) -> f32 {
+    steps: &[f64],
+    parameter_value: f64
+) -> f64 {
     let max_index = steps.len() - 1;
 
-    let index_float = parameter_value.max(0.0).min(1.0) * max_index as f32;
+    let index_float = parameter_value.max(0.0).min(1.0) * max_index as f64;
     let index_fract = index_float.fract();
 
     let index_low = index_float as usize;
@@ -47,9 +47,9 @@ pub fn map_parameter_value_to_value_with_steps(
 
 
 pub fn map_value_to_parameter_value_with_steps(
-    steps: &[f32],
-    internal_value: f32,
-) -> f32 {
+    steps: &[f64],
+    internal_value: f64,
+) -> f64 {
     let mut prev_step = *steps.first().expect("steps are empty");
 
     for (index, step) in steps[1..].iter().enumerate() {
@@ -57,11 +57,11 @@ pub fn map_value_to_parameter_value_with_steps(
 
         if internal_value <= step {
             let ratio = (internal_value - prev_step) / (step - prev_step);
-            let fraction = ((steps.len() - 1) as f32).recip();
+            let fraction = ((steps.len() - 1) as f64).recip();
 
             return ratio.mul_add(
                 fraction,
-                fraction * index as f32,
+                fraction * index as f64,
             );
         }
 
@@ -72,7 +72,7 @@ pub fn map_value_to_parameter_value_with_steps(
 }
 
 
-pub fn round_to_step(steps: &[f32], value: f32) -> f32 {
+pub fn round_to_step(steps: &[f64], value: f64) -> f64 {
     let mut prev_step = *steps.first().expect("steps are empty");
 
     for step in &steps[1..] {
@@ -104,7 +104,7 @@ mod tests {
 
     use super::*;
 
-    fn get_all_steps() -> Vec<f32> {
+    fn get_all_steps() -> Vec<f64> {
         let mut steps = Vec::new();
 
         steps.append(&mut OPERATOR_RATIO_STEPS.to_vec());
@@ -119,7 +119,7 @@ mod tests {
         steps
     }
 
-    fn valid_parameter_value(value: f32) -> bool {
+    fn valid_parameter_value(value: f64) -> bool {
         !(value.is_nan() || value > 1.0 || value < 0.0)
     }
 
@@ -164,7 +164,7 @@ mod tests {
 
     #[test]
     fn test_step_mapping(){
-        fn prop(value: f32) -> TestResult {
+        fn prop(value: f64) -> TestResult {
             if value < 0.0 || value > 1.0 {
                 return TestResult::discard();
             }
@@ -189,12 +189,12 @@ mod tests {
             TestResult::from_bool(inner == new_inner)
         }
 
-        quickcheck(prop as fn(f32) -> TestResult);
+        quickcheck(prop as fn(f64) -> TestResult);
     }
 
     #[test]
     fn test_map_value_to_parameter_value_with_steps_valid_number(){
-        fn prop(value: f32) -> TestResult {
+        fn prop(value: f64) -> TestResult {
             if value < 0.0 {
                 return TestResult::discard();
             }
@@ -207,7 +207,7 @@ mod tests {
             TestResult::from_bool(valid_parameter_value(value))
         }
 
-        quickcheck(prop as fn(f32) -> TestResult);
+        quickcheck(prop as fn(f64) -> TestResult);
     }
 
     #[test]
@@ -256,7 +256,7 @@ mod tests {
 
     #[test]
     fn test_smooth_step_mapping(){
-        fn prop(parameter_value: f32) -> TestResult {
+        fn prop(parameter_value: f64) -> TestResult {
             if parameter_value < 0.0 || parameter_value > 1.0 {
                 return TestResult::discard();
             }
@@ -285,7 +285,7 @@ mod tests {
             TestResult::from_bool(success)
         }
 
-        quickcheck(prop as fn(f32) -> TestResult);
+        quickcheck(prop as fn(f64) -> TestResult);
     }
 
     #[test]
